change session id to int64

+++ auth_test.go
@@ -252,7 +252,7 @@ func getDefaultSnowflakeConn() *snowflakeConn {
 }
 
 func TestUnitAuthenticateWithTokenAccessor(t *testing.T) {
-	expectedSessionID := 123
+	expectedSessionID := int64(123)
 	expectedMasterToken := "master_token"
 	expectedToken := "auth_token"
